move geojson-related to mapbox package
