+++ mapbox/build.gradle.kts
@@ -43,6 +43,9 @@ kotlin {
 
     sourceSets["commonMain"].dependencies {
         implementation(deps.kotlin.stdlibCommon)
+        implementation(deps.spatialk.geojson)
+        implementation(deps.spatialk.turf)
+        implementation(deps.spatialk.geojsonDsl)
     }
 
     sourceSets["commonTest"].dependencies {
