+++ mapbox/src/iosMain/kotlin/com/cuhacking/mapbox/GeoJsonSource.kt
@@ -1,4 +1,4 @@
-package com.cuhacking.atlas.common
+package com.cuhacking.mapbox
 
 import io.github.dellisd.spatialk.geojson.Feature
 import io.github.dellisd.spatialk.geojson.FeatureCollection
